Initial non-Linux changes
