+++ pkg/stanza/operator/input/journald/journald_nonlinux.go
@@ -1,0 +1,19 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build !linux
+// +build !linux
+
+package journald // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/journald"
+
+import (
+	"errors"
+
+	"go.uber.org/zap"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
+)
+
+func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
+	return nil, errors.New("journald input operator is only supported on linux")
+}
