+++ receiver/journaldreceiver/factory_test.go
@@ -4,9 +4,15 @@
 package journaldreceiver // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/journaldreceiver"
 
 import (
+	"context"
+	"runtime"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/consumer/consumertest"
+	"go.opentelemetry.io/collector/receiver/receivertest"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/journaldreceiver/internal/metadata"
 )
@@ -16,4 +22,31 @@ func TestNewFactory(t *testing.T) {
 		factory := NewFactory()
 		require.EqualValues(t, metadata.Type, factory.Type())
 	})
+}
+
+func TestCreateDefaultConfig(t *testing.T) {
+	factory := NewFactory()
+	cfg := factory.CreateDefaultConfig()
+	require.NotNil(t, cfg, "failed to create default config")
+}
+
+func TestCreateAndShutdown(t *testing.T) {
+	factory := NewFactory()
+	defaultConfig := factory.CreateDefaultConfig()
+	cfg := defaultConfig.(*JournaldConfig) // This cast should work on all platforms.
+
+	ctx := context.Background()
+	settings := receivertest.NewNopCreateSettings()
+	sink := new(consumertest.LogsSink)
+	receiver, err := factory.CreateLogsReceiver(ctx, settings, cfg, sink)
+
+	if runtime.GOOS == "linux" {
+		assert.NoError(t, err)
+		require.NotNil(t, receiver)
+		assert.NoError(t, receiver.Shutdown(ctx))
+	} else {
+		assert.Error(t, err)
+		assert.IsType(t, component.ErrDataTypeIsNotSupported, err)
+		assert.Nil(t, receiver)
+	}
 }
