+++ receiver/journaldreceiver/journald_nonlinux.go
@@ -14,8 +14,6 @@ import (
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/receiver"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/adapter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/journaldreceiver/internal/metadata"
 )
 
@@ -25,18 +23,6 @@ func newFactoryAdapter() receiver.Factory {
 		metadata.Type,
 		createDefaultConfig,
 		receiver.WithLogs(createLogsReceiver, metadata.LogsStability))
-}
-
-type JournaldConfig struct {
-	adapter.BaseConfig `mapstructure:",squash"`
-}
-
-func createDefaultConfig() component.Config {
-	return &JournaldConfig{
-		BaseConfig: adapter.BaseConfig{
-			Operators: []operator.Config{},
-		},
-	}
 }
 
 func createLogsReceiver(
